+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ## Fixed
+
 - Fix support for external fragments
 
 ## 1.0.0-beta.2 - 2022-03-27